Fix compile issue on Windows.

+++ Ecs/Graphics/include/NovelRT/Ecs/Graphics/SpriteRendererSystem.hpp
@@ -219,8 +219,8 @@ namespace NovelRT::Ecs::Graphics
                         _pendingStagingBuffers.push_back(
                             {_submittedUploads, textureStagingBuffer, gfxContext, uploadCmdList});
                         _textureCache.emplace(
-                            assetId, TextureCacheItem{texture2DRegion, Maths::GeoVector2F(textureMetadata.width,
-                                                                                          textureMetadata.height)});
+                            assetId, TextureCacheItem{texture2DRegion, Maths::GeoVector2F(static_cast<float>(textureMetadata.width),
+                                                                                          static_cast<float>(textureMetadata.height))});
                         _pendingLoads.erase(assetId);
                     });
             }
